+++ packages/personal/src/pursers_personal/__init__.py
@@ -2,7 +2,7 @@
 
 from importlib.metadata import PackageNotFoundError, version
 
-PRODUCT_VERSION = "5.0.0a16"
+PRODUCT_VERSION = "5.0.0a17"
 
 try:
     __version__ = version("pursers-personal")